Create quiescenceSearch.py
I implemented the Quiescence Search algorithm into the bot in order to avoid the "horizon effect", where the bot incorrectly evaluates positions due to the limited depth search. It works by extending the search depth to only captures, which ensures all volatile positions are played out until they are "stable", which makes sure that the bot returns an accurate move.